# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger for Fiber",
        "title": "Weather CRUD API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "zhan860127.gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:7800",
    "basePath": "/",
    "paths": {
        "/api/v1/del/{City}/{month}/{tempture}": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Weather"
                ],
                "summary": "delect the tempture in city on month",
                "parameters": [
                    {
                        "type": "string",
                        "description": "City",
                        "name": "City",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "month",
                        "name": "month",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功刪除",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "刪除失敗",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/modify/{City}/{month}/{tempture}": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Weather"
                ],
                "summary": "Modify the tempture in city on month",
                "parameters": [
                    {
                        "type": "string",
                        "description": "City",
                        "name": "City",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "month",
                        "name": "month",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "tempture",
                        "name": "tempture",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功修改",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "修改失敗",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/new/{City}/{month}/{tempture}": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Weather"
                ],
                "summary": "create the tempture in city on month",
                "parameters": [
                    {
                        "type": "string",
                        "description": "City",
                        "name": "City",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "month",
                        "name": "month",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "tempture",
                        "name": "tempture",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功新增",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "查詢失敗",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/weather/{City}/{month}": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Weather"
                ],
                "summary": "Get City_month Tempture",
                "parameters": [
                    {
                        "type": "string",
                        "description": "City",
                        "name": "City",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "month",
                        "name": "month",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查詢月份之城市溫度",
                        "schema": {
                            "type": "number"
                        }
                    },
                    "404": {
                        "description": "查詢失敗",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/encrpt/{message}": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Encrpt"
                ],
                "summary": "encrpt message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "message",
                        "name": "message",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功加密訊息",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}